mds: handle snap trimming when snapids aren't tied to a single SnapRealm
In the past, we only had a snapid in one place (its original root SnapRealm), so
detecting they'd been deleted was easy: we just zapped it in the
handle_client_snap_rm chain of functions. Now they might be disconnected, so
we have to be cleverer. For now, route it via the OSDMap and check whenever
we open the SnapRealm.

Signed-off-by: Greg Farnum <gfarnum@redhat.com>