TODO mds: hack up a better immediate-snap-deletion
I was accidentally relying solely on open_snaprealm() and the
OSDMap updates to delete snapids. That's clearly insufficient; so
hack up a way to delete from all open children of a SnapRealm when you
delete a snapid.

The real solution to this will need to be more complicated, and probably
bound up with not storing every active snapid in-memory on the SnapServer. Or
possibly just tracking deleted snapshots in the SnapClient and more aggressively
trimming stale snapids.

Signed-off-by: Greg Farnum <gfarnum@redhat.com>