client: convert ceph_ll_link to UserPerm and remove struct stat parameter
The main user of this API (ganesha) doesn't do anything with the
returned attributes, so there's no real point in returning them
there.

Also, we're not guaranteed to have any caps on the target inode
after the link operation, so in the case of FUSE (which does
require the post-op attributes) we should really do a getattr
to get the latest attributes.

Signed-off-by: Jeff Layton <jlayton@redhat.com>